requirements.txt
aiohttp==3.9.1
asyncio==3.4.3
pandas==2.1.4
tqdm==4.66.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0 